Carry fractional tempo values through the MNX and MusicXML exporters (#124)
* Map every note value MNX can express

`NoteType::Note256th` and `NoteType::Note4096th` had no case in the
`NoteValueBase` conversion even though mnxdom defines both. An unmapped
value falls through to the `default` and throws, so a 256th note
anywhere in a score failed the MNX export rather than converting.


* Take MNX tempos from the expression's playback settings

An MNX tempo is a playback instruction. It carries a number and a note
value and nothing else, with no way to give it visible text, so it should
state what Finale plays rather than what the page shows. It was taking a
metronome mark's displayed equation instead, which disagrees with the
playback settings whenever the two were configured independently. In
`metronome_marks.musx` the first mark displays an eighth while its
playback holds a quarter, so it exported at half the tempo Finale sounds.

The beat unit needs care. MNX counts one note value per minute, so the
unit has to be a base with some number of dots, and Finale's is a raw EDU
duration that need not be either. `calcDurationInfoFromEdu` answers with
the closest base and dot count for any duration in range rather than
refusing, so the answer is now spelled back out and compared before it is
trusted, and an unspellable unit is restated as a count of quarter notes.
That restatement is exact: MNX bpm is a real number as of schema version
34, and dividing by the EDUs in a quarter note divides by a power of two.
The same path catches a unit outside the range the helper accepts, which
previously threw.

Adds src/formats/mnx/design-decisions.md, which the format had no
equivalent of, recording both decisions.


* Carry fractional tempo values through both exporters

Updates the mnxdom pin to 698f8a0, which follows MNX schema version 34 in
widening `bpm` from an integer to a number, and the musxdom pin to
c870522, which reports a Tempo tool change as the tempo its stored ratio
encodes rather than rounding it to an integer.

Two consequences follow.

A metronome equation whose number is written with a decimal point, such
as "132.5", now classifies as a metronome mark and keeps its digits. The
parser accepted only digits before, so such a marking was not recognized
at all and degraded to plain tempo text. MusicXML `<per-minute>` is a
string and carries the number as printed; MNX takes it only where an
expression has no playback settings to prefer.

`<per-minute>` needs its own formatting. `std::to_string` on a double
writes 120.000000, and Denigma calls `setlocale(LC_ALL, "")`, so the
stream and printf conversions would emit the user's decimal separator and
write 132,5 into a file every reader parses as a number. The digits are
assembled from integer conversions instead. `std::to_chars` for floating
point would answer this but is unavailable at the minimum macOS
deployment target.

`musicXmlQuarterNotesPerMinute` now takes the tempo and beat unit
directly. Its caller for Tempo tool changes was building a `TempoInfo`,
whose `beatsPerMinute` is an `int`, purely to reach it, which truncated
the newly fractional value. The `TempoInfo` overload had one caller left
and is gone.


---------

+++ include/denigma/classify/expressions.h
@@ -149,8 +149,8 @@ struct MetronomeMark
     std::string noteGlyphName;
     /// @brief Number of augmentation dots following the note.
     std::size_t augmentationDots{};
-    /// @brief Number displayed to the right of the equals sign.
-    int displayedBeatsPerMinute{};
+    /// @brief Number displayed to the right of the equals sign. It may have a fractional part.
+    double displayedBeatsPerMinute{};
 };
 
 /// @struct TempoAlteration
